refactor code
In my last commit I discocered that I would not be able to do a post request due the data being nested. Therefore I had to refactor my data base to give it one end point, therefore I had to make some adjustments to my components to workwith an array of objects that were not nest.